Start assembling the amalgamation
My goal is to have a single .c file with no dependencies that you can drop into a project. The hard part of that is the
libsodium dependency. I am concatenating the necessary files from it into crypto.c, with slight tweaking to make it
compile.

Also, two files I forgot to add (rng.c/h) are included here.