Fix some warnings about unsigned long longs
The gcc options turned on warn about unsigned long longs, so they are now wrapped
in versions that take ints. (The code here is compiled with two separate sets of
options -- hyper-pedantic for the code unique to cses, and normal for the libsodium
code. The unsigned long longs are only on the libsodium side.)